requirements.txt
PyYAML==5.4.1
tensorboard==2.6.0
torch==1.11.0
torchvision==0.12.0
torchmetrics==0.5.1
tqdm==4.62.3
timm==0.4.12
torchtext==0.12.0
einops==0.4.0
vit-pytorch==0.26.5
protobuf==3.20.0
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git